Extract YearEndChartPrinter
