+++ src/main/java/chart/tasks/ChartOfChartsTask.java
@@ -2,7 +2,7 @@ package chart.tasks;
 
 import java.io.IOException;
 import java.sql.SQLException;
-import java.util.Iterator;
+import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
@@ -10,6 +10,7 @@ import chart.ChartConfig;
 import chart.ChartRun;
 import chart.Song;
 import chart.postgres.MultiChartLoader;
+import chart.postgres.YearEndChartPrinter;
 
 public class ChartOfChartsTask {
     public static void main(String[] args) throws IOException, SQLException, ClassNotFoundException {
@@ -23,14 +24,7 @@ public class ChartOfChartsTask {
         ChartConfig config = TaskUtils.getConfig();
         Map<Song, ChartRun> chartRuns = new MultiChartLoader(config).getAllChartRuns(year);
 
-        Iterator<ChartRun> chartOfCharts = chartRuns.values().stream().sorted().collect(Collectors.toList()).iterator();
-        for (int pos = 1; chartOfCharts.hasNext(); pos++) {
-            ChartRun chartRun = chartOfCharts.next();
-            print(pos, chartRun);
-        }
-    }
-
-    private static void print(int pos, ChartRun chartRun) {
-        System.out.println(String.format("%02d\t%s", pos, chartRun.toString()));
+        List<ChartRun> yec = chartRuns.values().stream().sorted().collect(Collectors.toList());
+        YearEndChartPrinter.printYearEndChart(yec);
     }
 }
